Refactor image resize as internal extension method
